+++ SimulationResponse.h
@@ -15,7 +15,8 @@
 #define SIMULATIONRESPONSE_H
 
 #include <string>
-#include "Listener.h"
+//#include "Listener.h"
+#include "Functor.h"
 
 /*!
  * Represents any possible response of a simulation. Any infrastructure or event the model can declare one of its own attribute as a simulation response. It just have to create a SimulationResponse object, passing the access to the method that gets the response value and including this SimulationResponse in the corresponding list of the model
