+++ nbproject/Makefile-variables.mk
@@ -7,13 +7,13 @@ CND_BASEDIR=`pwd`
 CND_BUILDDIR=build
 CND_DISTDIR=dist
 # Debug configuration
-CND_PLATFORM_Debug=GNU-Linux
-CND_ARTIFACT_DIR_Debug=dist/Debug/GNU-Linux
+CND_PLATFORM_Debug=GNU-MacOSX
+CND_ARTIFACT_DIR_Debug=dist/Debug/GNU-MacOSX
 CND_ARTIFACT_NAME_Debug=genesys-reborn
-CND_ARTIFACT_PATH_Debug=dist/Debug/GNU-Linux/genesys-reborn
-CND_PACKAGE_DIR_Debug=dist/Debug/GNU-Linux/package
+CND_ARTIFACT_PATH_Debug=dist/Debug/GNU-MacOSX/genesys-reborn
+CND_PACKAGE_DIR_Debug=dist/Debug/GNU-MacOSX/package
 CND_PACKAGE_NAME_Debug=genesys-reborn.tar
-CND_PACKAGE_PATH_Debug=dist/Debug/GNU-Linux/package/genesys-reborn.tar
+CND_PACKAGE_PATH_Debug=dist/Debug/GNU-MacOSX/package/genesys-reborn.tar
 # Release configuration
 CND_PLATFORM_Release=GNU-Linux
 CND_ARTIFACT_DIR_Release=dist/Release/GNU-Linux
